메인페이지 백엔트 key_value작업 완료

+++ parkingmate/src/main/java/com/app/parkingmate/domain/VO/UserVO.java
@@ -11,7 +11,7 @@ import java.util.Date;
 public class UserVO {
 //    회원 고유 id
 //    not null
-    private Integer UserId;
+    private Long UserId;
 //    회원 닉네임
 //    not null
     private String UserNickName;
